Add support for pathSuffix
This change adds support for a simple pathSuffix attribute in an app's
intent filter that can be applied to the path or ssp to match the end of
either. This enables filters that match based on file extensions as a
common use case.

This change additionally finally exposes the pathAdvancedPattern
attribute to enable more complex matching logic. This feature was added
back in API 26, but never exposed as a supported attribute in app
manifests, essentially making it useless to the average developer.

Test: atest IntentFilterTest
Bug: 139934742
Bug: 35193418

Change-Id: Ib58d5fb5419e621497e402b16a3a31c0c73f77cd