Check slot in reboot function
As part of the multi-client ror support, we want to compare the
intended boot slot from the client with the actual slot to boot
into.

Note boot control 1.2 is required to query the actual boot slot;
and we haven't installed V1.2 on Pixel3 and before. So the check is
skipped today if an old boot control HAL is installed on device.

Bug: 173808057
Test: atest CtsAppSecurityHostTestCases:ResumeOnRebootHostTest;
atest FrameworksServicesTests:RecoverySystemServiceTest

Change-Id: I13f9f5c75514424b70f96556b14a13783e9b6a7d