Receive consume time in InputPublisher
When the app reads an input event from the socket, it will now mark the
time at which it read the event. This time will later be provided to
InputDispatcher.

InputDispatcher can use this information to determine how long the
application has spent before the event was read. This step (socket read)
must occur before the app does any further frame processing. Therefore,
it will provide a better time bound for how long an app took to process
an input event.

InputDispatcher will report this socket latency as part of the
end-to-end touch latency metrics.

Bug: 169866723
Test: looked at the data printed locally to logcat
Change-Id: I14f77d3f9a4471b0be582a0852604cc9811e8d0a