Cast ApkAssets to uintptr_t before jlong
Ever since e40acd00457d7449f9814e9cde3614aa8141ad03 was submitted, we
have been getting reports of leaking of fds owned by ApkAssets.

The code that invokes the finalizer in NativeAllocationRegistry uses
the following code to cast the jlong free function to the pointer of the
native free function:
void* nativePtr = reinterpret_cast<void*>(static_cast<uintptr_t>(ptr));

In ApkAssets, we should be doing the reverse of these casts to give
NativeAllocationRegistry a pointer to a free function that will be
interpreted correctly.

Bug: 178571382
Test: manual
Change-Id: I6cd3f5cc9fd3693a1288e0309e9e49b08c0d5316