Add MAX/MIN_DEVICE_STATE constants to restrict possible states to range.
This adds a MAXIMUM_DEVICE_STATE and MINIMUM_DEVICE_STATE constant to
DeviceStateManager that contains the max allowed device state indentifier
which reserves any larger values for future defined constants in the
platform. The constant will be used by CTS to validate that the set
of supported states is in the range
[MINIMUM_DEVICE_STATE, MAXIMUM_DEVICE_STATE].

The values match the max allowed value returned by the default provider
impl which is defined in the XSD schema.

Bug: 159401801
Bug: 177235528
Test: atest DeviceStateTest
Change-Id: Ia5a8e2841bac19d36e9d6cd00950d09628a7f719