Create VibrationThread only after binder calling ID cleared
The creation of the VibrationThread does some setup for the WakeLock
that the thread must hold, which should only be executed after the
Binder calling ID has been cleared.

The necessary checks for this operation are already done by the method
enforceUpdateAppOpsStatsPermission, so the VibratorManagerService should
be able to setup the thread wakelock at that stage.

Fix: 180324695
Test: manual
Change-Id: I2f203fc4c7ba78687e9159af93b3d1acd9c826c9
(cherry picked from commit 7de460321e895541593b58952b2c02e13c549674)