Update domain verification Settings API
Adjusts the logic for selecting a domain, removing the selection for
other applications once a domain is verified or overridden by another
application being selected.

To this end, exposes a new Domain class with verified/selected booleans
to encapsulate the state, returned in the initial response for the
Settings screen API. This avoids the usage of the
getDomainVerificationSet API, so that can be restricted to only the
domain verification agent in a future change.

Also introduces the concept of a DomainOwner, which represents a single
package who has been granted any positive level of approval to open
a domain.

Bug: 177923646

CTS-Coverage-Bug: 179382047

Test: atest com.android.server.pm.verify.domain
Test: TODO with Settings changes

Change-Id: Ib14049e397154616f84a2264167ac30659cd81c9