Use stub-annotations to provide NonNull for hwbinder-stubs-docs
The annotations in srcs aren't doing anything because they are all
@hide.  The only thing keeping hwbinder-stubs-docs compiling is
a hack in metalava that treats anything with
hwbinder-stubs-docs-stubs.srcjar.rsp in its input files as
effectively having annotations_enabled: false.

Remove the unnecessry srcs, and add stub-annotations, which contains
a stub definition for android.annotation.NonNull, to hwbinder.stubs.
None of the rest of the annotations in srcs are used in the
generated stubs.

Test: m hwbinder.stubs
Change-Id: I7ec9afb41ea9617f25d1dffd3382fed90997b751